refactor: remove file fa-icons
- this file is only used in the zen-icon so it's no use to have it globally
