+++ src/components/zen-input/zen-input.stories.mdx
@@ -8,7 +8,6 @@ import data from '../../../stencilDocs.json';
 const compData = data.components.find(n => n.tag === 'zen-input');
 const argTypes = getArgTypes(compData);
 
-import { litHtmlIcon, styles } from '../helpers/fa-icons';
 import { faUser, faKey, faSearch } from '@fortawesome/pro-light-svg-icons';
 
 <Meta title="Forms/Input" component="zen-input" argTypes={argTypes} />
